halal-verify/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-telegram-bot==21.6
slowapi==0.1.9
limits==3.13.0
pytest==8.3.3
pytest-asyncio==0.24.0
pillow==10.4.0
pytesseract==0.3.13
